+++ src/components/reviews/review/review.js
@@ -1,15 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
 
 import Rate from '../../rate';
 import styles from './review.module.css';
+import { userSelector } from '../../../redux/selectors';
 
-const Review = ({ user, text, rating }) => (
+const Review = ({ user, text, rating, userId }) => (
   <div className={styles.review} data-id="review">
     <div className={styles.content}>
       <div>
         <h4 className={styles.name} data-id="review-user">
-          {user}
+          {user.name}
         </h4>
         <p className={styles.comment} data-id="review-text">
           {text}
@@ -23,7 +25,10 @@ const Review = ({ user, text, rating }) => (
 );
 
 Review.propTypes = {
-  user: PropTypes.string,
+  user: PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    name: PropTypes.string.isRequired,
+  }).isRequired,
   text: PropTypes.string,
   rating: PropTypes.number.isRequired,
 };
@@ -32,4 +37,6 @@ Review.defaultProps = {
   user: 'Anonymous',
 };
 
-export default Review;
+export default connect((state, props) => ({
+  user: userSelector(state, props),
+}))(Review);
